+++ runtests.py
@@ -15,10 +15,7 @@ if __name__ == '__main__':
     if not os.environ.get('SAMPLE_APP', False):
         args.insert(2, 'openwisp_users')
         args.insert(3, 'testapp.tests')
-        args.insert(4, 'testapp.test_multitenancy')
     else:
         args.insert(2, 'openwisp2')
         args.insert(3, 'testapp.tests')
-        args.insert(4, 'testapp.test_multitenancy')
-    # args.insert(5, '-v 3')
     execute_from_command_line(args)