+++ CHANGES.rst
@@ -9,7 +9,9 @@ WIP
 - `organization owners <https://github.com/openwisp/openwisp-users#organization-owners>`_:
   [admin] Added support for organization owners
 - [admin] Made organization owners read-only to non-superusers
-- [model] Added a helper to cache user permissions
+- `organization permissions <https://github.com/openwisp/openwisp-users#organization-permissions>`_:
+  [admin] Allowed administrator role to access organization admin
+- [model] Added `user permission helpers <https://github.com/openwisp/openwisp-users#permissions-helpers>`_
 
 Version 0.2.2 [2020-05-04]
 --------------------------